Complete decoupling of gas payment by removing the imbalance as a parameter of fill_gas and empty_unused_gas
Also:
* Rename GasHandling to GasHandler for consistency.
* Rename return_unused_gas to empty_unused_gas as a better name.
* Add default implementation of GasHandlers to the test which impl Trait.

Passing the imbalance resulted from withdrawing from an account to the refund function was just a trivial optimisation so during a contract we only address the imbalance once. However keeping that would cost us maintaining an ugly coupling between gas handling and gas payment. So we got rid of that. between.