Cherry-pick substrate commit #54e6298523ea896f2129546c26dd7a47dbbeedc9 which will fix transaction payment runtime api (#6792 issue on substrate)
The transaction payment runtime api used its own extrinsic generic parameter. This is wrong, because this resulted in using always the native extrinsic. If there was a runtime upgrade that changed the extrinsic in some way, it would result in the api breaking. The correct way is to use the `Extrinsic` from the `Block` parameter. This is on the node side the opaque extrinsic and on the runtime side the real extrinsic.